Move hooks from /src/ui/hooks to /src/hooks
