+++ src/hooks/useStaticAsset.js
