+++ src/ui/states/approveTransaction/details/DocumentsBatchDetails.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import { useSdk } from '../../../hooks/useSdk'
+import { useSdk } from '../../../../hooks/useSdk'
 import './documents.batch.details.css'
 import DocumentCreateTransition from './DocumentCreateTransition'
